\name{CGene-package}
\alias{CGene-package}
\alias{CGene}
\docType{package}
\title{
Causal Genetic Analysis
}
\description{
Implements statistical techniques to understand the causal pathways between genetic markers and a primary outcome when an intermediate phenotype is also associated with both the marker the the primary outcome with population-based data.
}
\details{
\tabular{ll}{
Package: \tab CGene\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2010-12-22\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
The functions provided allow the investigator to test if a genetic marker is associated with a primary outcome through pathways other than that of an intermediate secondary phenotype. The functions allow for the primary outcome to be (1) continuous, (2) discrete, (3) survival modeled parametrically, or (4) survival modeled semi-parametrically
}
\author{
Peter Lipman

Maintainer: Contact <pjlipman@fas.harvard.edu>
}
\references{
Vansteelandt S. et al: On the Adjustment for Covariates in Genetic Association Studies: A Novel, Simple Principle to Infer Direct Causal Effects. Genetic Epidemiology 2009; 33(5): 394-405.
Lipman P. et al: Inferring Genetic Causal Effects on Survival Data with Associated Endo-Phenotypes. Genetic Epidemiology 2011 (in press).
}

